dailyfresh/requirement.txt
amqp==2.5.2
billiard==3.5.0.5
celery==4.1.1
certifi==2019.11.28
chardet==3.0.4
Django==1.8.2
django-redis==4.8.0
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.8
importlib-metadata==1.3.0
itsdangerous==0.24
kombu==4.6.7
more-itertools==8.0.2
mutagen==1.43.0
olefile==0.46
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.3
redis==3.2.0
requests==2.22.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
